Remove warnings when compiling current files

+++ main.h
@@ -1,4 +1,4 @@
 void get_boost_version ();
-void print_line (int symbols_count, char symbol);
+void print_line (size_t symbols_count, char symbol);
 void print_line (char symbol);
 void get_help ();
